requirements.txt
django~=4.2
python-decouple==3.4
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
psycopg2==2.9.9
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
drf-yasg[validation]

# Code quality and linter
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Test
pytest==8.1.1
pytest-django==4.8.0
